third_party/WebKit/LayoutTests/crypto/digest.html

<!DOCTYPE html>
<html>
<head>
<script src="../fast/js/resources/js-test-pre.js"></script>
<script src="resources/common.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>

<script>
description("Tests cypto.subtle.digest.");

jsTestIsAsync = true;

function printRejectedResult(value)
{
    debug("    rejected with value of " + value);
}

function printAcceptedResult(result)
{
    debug("    = " + byteArrayToHexString(new Uint8Array(result)));
}

Promise.resolve(null).then(function() {
    debug("SHA1 of []");
    return crypto.subtle.digest({name: 'sha-1'}, new Uint8Array([]));
}).then(function(result) {
    printAcceptedResult(result);

    debug("SHA1 of [0x0]")
    return crypto.subtle.digest({name: 'sha-1'}, new Uint8Array([0]));
}).then(function(result) {
    printAcceptedResult(result);

    // Pass invalid data to digeset()
    shouldThrow("crypto.subtle.digest({name: 'sha-1'})");
    shouldThrow("crypto.subtle.digest({name: 'sha-1'}, null)");
    shouldThrow("crypto.subtle.digest({name: 'sha-1'}, 10)");

    // Pass invalid algorithmIdentifiers to digest()
    data = new Uint8Array([0]);
    shouldThrow("crypto.subtle.digest({name: 'sha'}, data)");
    shouldThrow("crypto.subtle.digest(null, data)");
    shouldThrow("crypto.subtle.digest({}, data)");
}).then(finishJSTest, failAndFinishJSTest);

</script>

<script src="../fast/js/resources/js-test-post.js"></script>
</body>
</html>
