third_party/WebKit/LayoutTests/crypto/digest-expected.txt

Tests cypto.subtle.digest.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


SHA1 of []
    = [da 39 a3 ee 5e 6b 4b 0d 32 55 bf ef 95 60 18 90 af d8 07 09]
SHA1 of [0x0]
    = [5b a9 3c 9d b0 cf f9 3f 52 b5 21 d7 42 0e 43 f6 ed a2 78 4f]
PASS crypto.subtle.digest({name: 'sha-1'}) threw exception TypeError: Failed to execute 'digest' on 'SubtleCrypto': 2 arguments required, but only 1 present..
PASS crypto.subtle.digest({name: 'sha-1'}, null) threw exception TypeError: Invalid dataBuffer argument.
PASS crypto.subtle.digest({name: 'sha-1'}, 10) threw exception TypeError: Invalid dataBuffer argument.
PASS crypto.subtle.digest({name: 'sha'}, data) threw exception NotSupportedError: Algorithm: Unrecognized algorithm name.
PASS crypto.subtle.digest(null, data) threw exception TypeError: Algorithm: Not an object.
PASS crypto.subtle.digest({}, data) threw exception TypeError: Algorithm: name: Missing or not a string.
PASS successfullyParsed is true

TEST COMPLETE
